The following public method has been added:
DateTime GetThisOrNextDateTimeWhenTraded(
	DateTime startingDateTime ,
	Time dailyTime ,
	DateTime lastAcceptableDateTime )

Starting from startingDateTime, this method searches for the next dateTime when the benchmark is traded at time dailyTime.
The returned value is greater than or equal to startingDateTime, that is, startingDateTime itself is a valid candidate.
If no solution is found within maxNumberOfDaysToLookAheadFor days, then an exception is returned