The read only property public double[] Residuals has been added.
The read only property
public double CenteredTotalSumOfSquares
has been added

IMPORTANT: the read only property
public double PredictedResidualsSumOfSquares
has been added.
It is the PRESS statistic, a statistic based on the leave-one-out technique. This statistic is much more robust than Rsquare, to outliers.

The read only property
public double PredictedCenteredTotalSumOfSquares
has been added.

IMPORTANT: the read only property
public double CenteredPSquare
has been added.
It is a normalized version of the the PRESS statistics, and it can be used as a fitness alternative the CenteredRSquare: the CenteredPSquare is much more robust with respect to outliers.